Disable radeonSoftwareVersion field until needed driver update is public
The radeonSoftwareVersion field is not currently working correctly. This
update disables it while we wait for the necessary driver updates.

This update also changes WindowsTargetPlatformVersion for Visual Studio
2015 to target the Windows 8.1 SDK. The Windows 10 SDK is not installed
by default with Visual Studio 2015, and since AGS is based on DirectX 11
(and not DirectX 12), it does not need the Windows 10 SDK. The 8.1 SDK
is installed by default, so just target it, to avoid adding an
unnecessary dependency on an optional component of Visual Studio.