demo: Add a stripped down version of miniz v1.5
This adds miniz v1.5 with removed zlib API bits and other configuration
options, leaving just the basic portable inflate/deflate - this reduces
the library to around a third of the original size.

It will only be used in demo code to gauge the post-deflate as well as
pre-deflate efficiency of vertex/index codecs.