vertexcodec: Compute vertex block size dynamically
For decoding performance we'd like to do operations on the entire block
at once. For this to work the entire block should fit into a fixed size
scratch buffer if we don't want to do dynamic memory allocations.

We now compute the block size as min(8192 / vertex_size, 256); for
vertices up to 32 bytes this still uses size=256.