vertexcodec: Remove parallelogram prediction
This is not an easy decision but this change pulls out pgram predictor
from vertex codec.

The benefits of pgram predictor remain underwhelming. The largest
benefit on the data set seems to be ~8bpv which is pretty good, but the
usual benefit is around 4bpv (which is <1b per component!).

Conversely, pgram prediction adds a large amount of extra complexity (it
makes adjustments to index codec somewhat more cumbersome and requires
several alternate code paths in the vertex codec), and has a high decode
time penalty, more than doubling the decoding time for 16b vertex.

Moreover, in presence of multiple vertex streams, the cost increases
even further since we need to decode the index buffer multiple times, or
complicate the decoder even further.

Compared to something like replacing normal storage with octahedron
encoding which effortlessly saves 8 bpv, this is just not worth it. For
now we'll focus on pure vertex encoding performance.