vertexcodec: Make decoder memory-safe
We now should have all the appropriate bounds checks to make the decoder
memory-safe in the face of untrusted data; this has not been checked
with a fuzzer yet.

Additionally, this change tightens the estimates for encoded size to
match the worst case without overallocating; the resulting size is close
to vertex_count * vertex_size for large vertex_count values.