vertexcodec: Automatically enable SIMD at compile-time
This change refactors the SIMD defines to enable SIMD implementation
automatically when possible.

For NEON, just compiling for ARM should enable it by default.
For SSE, this requires /arch:AVX (for MSVC) or -mssse3 (for GCC/Clang).

MSVC doesn't seem to have /arch:SSSE3 switch; in general this is not
necessarily the right way to configure this - instead we should probably
compile both SIMD and scalar versions in and use cpuid to switch - but
this is good enough for now.