demo: Abbreviate simplify output
Instead of dumping per-LOD stats, we just print ACMR and Overfetch for
first and last LODs and cumulative codec stats. Overdraw tends to not
change much for LOD levels.

It would be interesting to print some sort of similarity metric as well
but it's not clear how to efficiently compute it...