indexgenerator: Optimize generateVertexRemap
We used to store an old->new id mapping in a hash table, using old index
as a key (with hash function/comparator working on vertex data). Then
we'd fill remap table based on this data. However, remap table *is* a
mapping from old->new ids so this is redundant.

Instead what we can do is use a hash set that just stores the old index
of a "canonical" vertex for each vertex - if vertices a & b are equal,
we'd only store one of them in the hash set and looking up either would
return that one.

This reduces the hash table size 2x which makes hash lookups faster; on
large meshes this results in ~10% speedup for remap generation.