simplify: Compute position remap table
Implement position remap table generation, that maps from a vertex index
to the vertex index with the same position (using binary equivalence).

This is very similar in operation to meshopt_generateVertexRemap, but
instead of generating a sequential mapping it maps some vertices to
others. This might make it easier to adapt the rest of the code to use
position remap...