simplify: Implement vertex classification
We now classify vertices into 4 kinds:

manifold - a vertex with unique position and manifold edges
border - a vertex with unique position that has 2 open edges
seam - a vertex with non-unique position that has 2 seam edges
locked - all other vertices

The classification is somewhat bare-bones - it doesn't properly account
for all cases and might have bugs. We will refine this as we go.