simplify: Implement manifold->locked edge collapse
Our general manifold collapse path works perfectly fine as long as the
source vertex is manifold (modulo a few asserts we don't really need).
This change reworks the seam collapse to assume seam->seam, and uses
manifold->X for everything else. This allows us to use this for
manifold->locked as well which this change enables.