simplify: Assert that collapse passes reduce triangle count
With new collapse logic each edge we collapse *must* result in at least
one triangle being removed; if it doesn't happen this means some logic
somewhere has a bug. No meshes in the test set trigger this, so change
if() to assert().

Also rename vertex_locked to collapse_locked as an indication that the
lock flags are specific to each collapse pass.