vertexcodec: Optimize lookup table generation
We run LUT generation in a static ctor right now; at some point it might
be interesting to remove the only static ctor the library has and do
lazy initialization. For this to work, the function has to be fast...

Before this change, it spent ~12K cycles to generate a 4K lookup table;
a lot of this was spent on branch mispredictions inside the loop. We now
use branchless code instead, which reduces the time to ~6K cycles.

Also verify that static construction works by checking the boolean...
This can fire if vertex buffer decoder is used before main().