simplify: Clean up quadric error metrics
Based on empirical test results, the weighting scheme described in
Garland99 works best:

- For triangle quadrics, we use triangle area as a weight; we currently
don't divide it by 3 to distribute it equally among vertices although
maybe we should...

- For edge quadrics, we use squared edge length as a weight; squaring
the length makes sure the dimension of quadric error is d^4 for edges as
well as faces.

Note that for edge quadrics, we end up using both edge and plane
quadrics which effectively uses a metric that's close to squared
distance-to-edge without any extra work.