tools/lodviewer: Improve energy efficiency
This change makes sure we only redraw the window when necessary. To make
this work well, we explicitly set the redraw flag when necessary, and
change the main loop to use glfwWaitEvents instead of glfwPollEvents to
avoid spending the cycles for event polling.

Note that manually setting redraw to true when necessary is important to
make sure we don't redraw the window on each mouse movement.