simplify: Optimize quadricError by reducing ALU pressure
We reassociate the computation to make sure that multiplications aren't
performed redundantly, which gets us from ~31 FP ALU to ~22.

This results in a modest - 3%? - performance improvement on some meshes.
Notably, some large meshes regress wrt performance because the new
metric generates fewer zeroes, which for meshes that end up with
edge_collapse_goal = 0 results in more passes. This needs to be
addressed separately.