simplify: Use loop[] instead of adjacency for edge quadrics
Instead of validating whether the edges are open when adding quadrics to
keep vertices close to the original edge, we now use the loop[] data.

This is identical to the existing implementation in terms of behavior,
but it's a bit faster since we don't need to search for the edge;
additionally, since loop[] is updated throughout the simplification, we
can now implement memoryless simplification by reinitializing quadric
weights before each pass.