simplify: Cleanup remaining TODOs
This change removes the rest of the TODO items. Specifically:

- We kinda give up on the border/seam->locked collapses. Border->locked
isn't that interesting; seam->locked would be, but it's highly
problematic - we currently use wedge metadata to find the target vertex
for a seam collapse; when collapsing onto a locked vertex, we can't use
wedge metadata since there's more than 2 wedges, we can't use adjacency
since it isn't updated between passes, and we can't use loop[] because
it tracks half edges so if edges point in the wrong direction we won't
be able to find the vertex.

- The pass error tracking code was reworked to find the last collapse
that actually happened and grab the error from here, which nicely
decouples TRACE-only code further

- Debugging output for lodviewer was put under TRACE - you now need
TRACE to build lodviewer but we were basically always using that anyway.

This change also cleans up edge handling a tiny bit - we only need to
check loop[] in one direction when picking collapses since it stores
half edges and also we now use the same logic for adding edge weights.

We're not really *done* with this algorithm - still needs a bit more
work on quadric errors and we need to figure out what to do with faceted
models, plus hairball.obj exhibits pathological behavior with too many
passes - but we're in a pretty solid shape at this point.