meshlets: Add an experimental version of meshlet builder
Assuming the input triangle order is optimized for vertex cache
locality, the algorithm greedily builds meshlets from the input data
according to the max vertex/triangle restrictions.

The output is performed using a struct with a hardcoded maximum for both
limits (64 for number of vertices and 126 for number of triangles),
which seems reasonable given the limitations of GPU hardware.

This algorithm can be used for cluster culling as well - the output is
somewhat more compact than a traditional index buffer, and one can
generate cone data per meshlet - although other split strategies might
be more appropriate.

The output of this algorithm is reasonable wrt meshlet stats, it remains
to be seen if further gains are possible.