Introduce MESHOPTIMIZER_EXPERIMENTAL tag
In addition to just noting the exprerimental fashion of some APIs in the
comment, more explicitly annotate the function definition.

This is to try to make sure experimental APIs aren't used in production
code as they are subject to change without notice or consideration for
backwards compatibility, and can have severe stability or performance
issues.