simplify: Remove incorrect comment
I confused myself by thinking that the interface was wrong when writing
this, but the interface is actually correct.

Simplifier doesn't read the vertex data besides position currently - it
uses index buffer generated over the entire vertex to infer attribute
seams. This makes it impossible to analyze seams separately for separate
attributes, e.g. to ignore normal seams, but it does work otherwise.