clusterizer: Added experimental cone generator
This is a sketch of an algorithm to generate a cone that can be used to
do backface culling. It computes normal cone for all triangles in a
meshlet/cluster and prepares the data for fast culling, with the demo
program computing rejection and trivial acceptance metrics.

Note that the current implementation has two issues:

- The cone construction is pretty sloppy - in many cases we can make the
resulting cone tighter which improves rejection rates.

- The cone culling works correctly for orthographic camera but doesn't
work for perspective camera. It seems that for perspective camera we
need to compute a cone in a different fashion, similarly to GeometryFX
article.