clusterizer: Pass meshlet by value to computeMeshletBounds
In C there's no support for references so we need to either pass meshlet
by value or by constant pointer. This change switches to pass by value
as it's cleaner for the callers.

This does mean that the caller has to copy a relatively large meshlet
structure, but the cost of copying the indices and - more importantly -
computing the bounding spheres and apex is substantially larger than a
memcpy.