indexgenerator: Introduce shadow IB generator
This function can build a "shadow" index buffer - an index buffer that
can be used alongside the main index buffer/vertex buffer to render a
mesh using only some of the attributes. Commonly you'd generate an index
buffer for just vertex positions and use that for Z-prepass or shadowmap
rendering, hence the name.

The shadow index buffer references the same indices but eliminates
UV/normal creases which improves the vertex reuse and can improve vertex
shading time by 10-20% depending on the mesh.