clusterizer: Fix zero-triangle clusters
computeBoundingSphere isn't safe to use for zero-triangle clusters since
it reads uninitialized memory. Additionally, it would be nice to have
zero-triangle clusters be trivially rejected instead of trivially
accepted - so we fill bounds with zeroes instead of setting the cutoff
to 1 for these.

This also requires adjusting the cone test to use non-strict equality,
hence the comment adjustments.