demo: Add pirate demo
This adds a demo of using OptMeshLoader for the same pirate.obj model
that we already have in the repository.

Also move OptMeshLoader to demo; I don't really want to commit to optmesh
format being anything other than an example so it'd be more appropriate to
treat js/ as the library proper, and OptMeshLoader as an example.