tools: Improve position/uv packing in meshencoder
Instead of encoding position/uv data into top bits, encode them into
bottom bits - that way vertex codec is more efficient, because the
deltas of the high byte are smaller.

We still use stock THREE.js materials so we now use mesh position to
offset the mesh and texture matrix to transform the UV coordinates.

For UV, we use a bounding rectangle which means we now support tiled
textures; for position, we still use uniform scale because the position
matrix will be used to transform the normal and we need to avoid
distorting the normal.

nyra.optmesh is reduced from 217242 bytes to 198078 bytes as a result.