Add generateVertexRemapMulti and generateShadowIndexBufferMulti
These algorithms have the same purpose as generateVertexRemap and
generateShadowIndexBuffer respectively, but they work on deinterleaved
vertices and take a set of vertex streams as an input instead of just
one.

All other algorithms were already usable on deinterleaved geometry
(sometimes via multiple calls to e.g. meshopt_remapVertexBuffer or
meshopt_encodeVertexBuffer), so these two are the last remaining.

Note that we still have a separate implementation for single-stream
remap. Unfortunately it looks like gcc is ~10% slower when using the
code for multiple streams with one stream - further performance analysis
is necessary to figure out if we want to remove the duplicate
implementations.